requirements.txt
# Core framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# GitHub App authentication
PyJWT==2.8.0
cryptography==42.0.2

# HTTP client (async)
httpx==0.27.0

# AI inference (CPU-only)
llama-cpp-python==0.2.56

# Structured logging
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
